+++ scraper.go
@@ -6,10 +6,13 @@ import (
 	"net/url"
 	"os"
 	"slices"
+	"strconv"
 	"strings"
 	"sync"
 
 	"github.com/common-nighthawk/go-figure"
+	"github.com/criticalsession/4scraper/config"
+	"github.com/criticalsession/4scraper/search"
 	"github.com/gocolly/colly"
 	"github.com/k0kubun/go-ansi"
 	"github.com/mitchellh/colorstring"
@@ -22,25 +25,93 @@ type DownloadableFile struct {
 }
 
 func main() {
-	// setup
-	silent, outDir, url := ParseFlags()
-	conf := ReadConfig()
+	flags := config.ParseFlags()
+	conf := config.ReadConfig()
 
-	// head
-	if !silent {
+	if flags.IsSearch {
 		figure.NewFigure("4scraper", "rectangles", true).Print()
-		fmt.Println(v)
-		fmt.Println("https://github.com/criticalsession/4scraper")
+		fmt.Printf("%s [%s]\n", config.Version, "Search Mode")
+		colorstring.Println("[dim]https://github.com/criticalsession/4scraper")
 		fmt.Println("")
 
-		if url == "" {
-			fmt.Print("> Enter thread url: ")
-			fmt.Fscan(os.Stdin, &url)
+		threads, err := search.FindInBoard(flags.SearchBoard, flags.SearchTerms)
+		if err != nil {
+			logErr(err)
+			return
+		}
+
+		if len(threads) > 0 {
+			colorstring.Printf("Search results for [green]%s[default] in [green]/%s/[default]\n\n", strings.Join(flags.SearchTerms, ", "), flags.SearchBoard)
+			for i, t := range threads {
+				colorstring.Printf("[bold][green][%d][default] https://boards.4chan.org/%s/thread/%d (Images: %d)\n", i+1, flags.SearchBoard, t.No, t.Images+1)
+				if t.Sub != "" {
+					fmt.Printf("  %s\n", t.Sub)
+				}
+
+				if t.Com != "" {
+					fmt.Printf("  %s\n", t.Com)
+				}
+
+				fmt.Println()
+			}
+
+			var downOption string
+			colorstring.Print("> Enter [green][number][default] to download thread, [green]all[default] to download all, [green]q[default] to quit: ")
+			fmt.Fscan(os.Stdin, &downOption)
+
+			downOption = strings.ToLower(downOption)
+
+			threadId, err := strconv.ParseInt(downOption, 10, 32)
+			if err != nil {
+				if downOption == "all" || downOption == "a" {
+					for _, t := range threads {
+						url := fmt.Sprintf("https://boards.4chan.org/%s/thread/%d", flags.SearchBoard, t.No)
+						colorstring.Printf("\nDownloading thread [green]%d[default]...\n", t.No)
+						downloadUrl(url, flags, conf)
+					}
+				} else {
+					return
+				}
+			} else {
+				threadId -= 1
+				if threadId >= 0 && int(threadId) < len(threads) {
+					colorstring.Printf("\nDownloading thread [green]%d[default]...\n", threads[threadId].No)
+					url := fmt.Sprintf("https://boards.4chan.org/%s/thread/%d", flags.SearchBoard, threads[threadId].No)
+					downloadUrl(url, flags, conf)
+				} else {
+					logErr(errors.New("invalid thread number"))
+				}
+			}
 		} else {
-			fmt.Println("Using arg url:", url)
+			fmt.Printf("No results found for %s in /%s/\n\n", strings.Join(flags.SearchTerms, ", "), flags.SearchBoard)
+			return
+		}
+	} else {
+		url := flags.Url
+
+		// head
+		if !flags.Silent {
+			figure.NewFigure("4scraper", "rectangles", true).Print()
+			fmt.Printf("%s [%s]\n", config.Version, "Download Mode")
+			fmt.Println("https://github.com/criticalsession/4scraper")
+			fmt.Println("")
+
+			if url == "" {
+				fmt.Print("> Enter thread url: ")
+				fmt.Fscan(os.Stdin, &url)
+			} else {
+				fmt.Println("Using arg url:", url)
+			}
+		}
+
+		downloadUrl(url, flags, conf)
+		if !flags.Silent {
+			fmt.Println("\nDownload finished. Thank you for using 4scraper!")
 		}
 	}
+}
 
+func downloadUrl(url string, flags config.FlagSettings, conf config.Config) {
 	// build download dir
 	board, threadId, err := extractBoardAndThreadId(url)
 	if err != nil {
@@ -49,8 +120,8 @@ func main() {
 	}
 
 	var downloadDir string
-	if len(outDir) > 0 {
-		downloadDir = outDir
+	if len(flags.OutDir) > 0 {
+		downloadDir = flags.OutDir
 	} else {
 		downloadDir = "downloads"
 		if conf.BoardDir {
@@ -80,14 +151,14 @@ func main() {
 	c.OnHTML(".fileText > a", func(h *colly.HTMLElement) {
 		addLinkToDownloadableFiles(h, &conf.Extensions, &files)
 
-		if !silent {
+		if !flags.Silent {
 			bar.Add(1)
 		}
 	})
 
 	c.Visit(url)
 
-	if len(files) == 0 && !silent {
+	if len(files) == 0 && !flags.Silent {
 		fmt.Println("No files found to download.")
 		return
 	}
@@ -117,7 +188,7 @@ func main() {
 
 	for i, file := range files {
 		if !conf.ParallelDownload {
-			if err := DownloadFile(file.Url, downloadDir, file.FileName, conf.UseOriginalFilename, silent, bar); err != nil {
+			if err := DownloadFile(file.Url, downloadDir, file.FileName, conf.UseOriginalFilename, flags.Silent, bar); err != nil {
 				logErr(err)
 			}
 		} else {
@@ -125,7 +196,7 @@ func main() {
 			queue <- true
 			go func(file DownloadableFile, i int) {
 				defer wg.Done()
-				if err := DownloadFile(file.Url, downloadDir, file.FileName, conf.UseOriginalFilename, silent, bar); err != nil {
+				if err := DownloadFile(file.Url, downloadDir, file.FileName, conf.UseOriginalFilename, flags.Silent, bar); err != nil {
 					errs = append(errs, err)
 				}
 
@@ -140,10 +211,6 @@ func main() {
 		for _, err := range errs {
 			logErr(err)
 		}
-	}
-
-	if !silent {
-		fmt.Println("\nDownload finished. Thank you for using 4scraper!")
 	}
 }